# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Drive API
  description: |
    A secure file sharing platform that allows users to upload, expose, and manage access to their files.
    
    ## Features
    - User authentication with phone numbers
    - File upload and management
    - Secure file exposure with time limits
    - Access request and approval system
    - IP-based access control and blacklisting
    - Premium phone-based sharing with secure tokens
    
    ## User Types
    - **Free Users**: Can expose one file at a time for maximum 5 minutes
    - **Premium Users**: Can expose multiple files with custom expiration times up to 24 hours, phone-based sharing
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain tokens through the login endpoint.
    
  version: 1.0.0
  contact:
    name: Drive API Support
    url: https://expose.tech/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.expose.drive.tech
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /premium/expose:
    post:
      summary: Create premium exposure
      description: Create file exposure with phone-based sharing for premium users
      tags: [Premium Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_id, permission, shared_with_phones]
              properties:
                file_id:
                  type: integer
                  description: ID of file to expose
                  example: 1
                permission:
                  type: string
                  enum: [read_only, download]
                  description: Permission level
                  example: "download"
                expires_minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  description: Expiration time in minutes (1-1440)
                  example: 720
                shared_with_phones:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  description: Phone numbers to share with
                  example: ["+918043652258", "+919876543210"]
                max_downloads_per_recipient:
                  type: integer
                  minimum: 1
                  description: Maximum downloads per recipient
                  example: 3
      responses:
        '201':
          description: Premium exposure created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PremiumExposureResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Premium features require premium subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /premium/exposures:
    get:
      summary: List premium exposures
      description: List all premium exposures for current user
      tags: [Premium Features]
      responses:
        '200':
          description: List of premium exposures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PremiumExposureResponse'
        '403':
          description: Premium features require premium subscription
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /premium/share/add:
    post:
      summary: Add recipient to exposure
      description: Add a new recipient to an existing premium exposure
      tags: [Premium Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exposure_hash, phone]
              properties:
                exposure_hash:
                  type: string
                  description: Hash of exposure to add recipient to
                  example: "abc123def456..."
                phone:
                  type: string
                  description: Phone number to share with
                  example: "+919999999999"
                max_downloads:
                  type: integer
                  minimum: 1
                  description: Maximum downloads for this recipient
                  example: 2
      responses:
        '201':
          description: Recipient added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PremiumShareResponse'
        '400':
          description: Invalid phone number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Premium exposure not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Phone number already has access to this file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /premium/share/revoke:
    post:
      summary: Revoke share access
      description: Revoke access for a specific recipient
      tags: [Premium Features]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [share_id]
              properties:
                share_id:
                  type: integer
                  description: ID of share to revoke
                  example: 1
      responses:
        '200':
          description: Share access revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Share access revoked successfully"
                  share_id:
                    type: integer
                  revoked_phone:
                    type: string
        '404':
          description: Share not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to revoke this share
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /premium/access/{access_token}:
    get:
      summary: Access premium file
      description: Access premium shared file with token and phone verification
      tags: [Premium Features]
      security: []
      parameters:
        - name: access_token
          in: path
          required: true
          description: Access token for the shared file
          schema:
            type: string
        - name: phone
          in: query
          required: true
          description: Phone number for verification
          schema:
            type: string
            example: "+918043652258"
      responses:
        '200':
          description: File access information
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_name:
                    type: string
                    example: "document.pdf"
                  file_size:
                    type: integer
                    example: 1048576
                  mime_type:
                    type: string
                    example: "application/pdf"
                  permission:
                    type: string
                    enum: [read_only, download]
                  download_count:
                    type: integer
                    description: Number of downloads used by this recipient
                  max_downloads:
                    type: integer
                    nullable: true
                    description: Maximum downloads allowed for this recipient
                  expires_at:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          description: Invalid access token format or phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid access token or phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Access token has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /premium/download/{access_token}:
    get:
      summary: Download premium file
      description: |
        Download premium shared file with token and phone verification.
        
        **Note**: This endpoint returns binary content and should not be tested directly in Swagger UI.
        Use a proper HTTP client or browser for file downloads.
      tags: [Premium Features]
      security: []
      parameters:
        - name: access_token
          in: path
          required: true
          description: Access token for the shared file
          schema:
            type: string
        - name: phone
          in: query
          required: true
          description: Phone number for verification
          schema:
            type: string
            example: "+918043652258"
      responses:
        '200':
          description: File download (binary content)
          content:
            '*/*':
              schema:
                type: string
                format: binary
                description: Binary file content
          headers:
            Content-Disposition:
              description: File attachment with original filename
              schema:
                type: string
            Content-Type:
              description: MIME type of the file
              schema:
                type: string
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
        '400':
          description: Invalid access token format or phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Download limit exceeded or download not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid access token, phone number, or file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Approval Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "File not found"
        code:
          type: string
          description: Error code
      required:
        - detail
...
    PremiumShareResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        exposure_id:
          type: integer
          example: 1
        shared_with_phone:
          type: string
          example: "+918043652258"
        access_token:
          type: string
          description: Secure access token for this recipient
          example: "def789ghi012..."
        status:
          type: string
          enum: [active, expired, revoked]
          example: "active"
        download_count:
          type: integer
          description: Number of downloads used by this recipient
          example: 2
        max_downloads:
          type: integer
          nullable: true
          description: Maximum downloads allowed for this recipient
          example: 5
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-22T22:30:00Z"
        created_at:
          type: string
          format: date-time
          example: "2025-01-22T10:30:00Z"
    PremiumExposureResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        exposure_hash:
          type: string
          description: Unique hash for the exposure
          example: "abc123def456..."
        file_id:
          type: integer
          example: 1
        permission:
          type: string
          enum: [read_only, download]
          example: "download"
        exposure_type:
          type: string
          example: "premium"
        premium_token:
          type: string
          nullable: true
          description: Main premium token for the exposure
          example: "xyz789uvw012..."
        expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-22T22:30:00Z"
        shared_recipients:
          type: array
          items:
            $ref: '#/components/schemas/PremiumShareResponse'
          description: List of recipients and their access details
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-01-22T10:30:00Z"
...